[cleanup] move more classes to subfolders
