+++ utils/entitydescriptor.cpp
