+++ utils/entitydescriptor.h
